Unbundled lodash and moved to a minified file.
This is a fairly substantial change of how the project is deployed
to the outside, and includes adding ES2015-module awareness to
outside consumers. (Presumably mostly other projects of mine at
this point.) This change was done partly with view to not having
any sort of copyright questions regarding the build, for automatic
CDN stuff.

It does mean, however, that anyone using it needs to get lodash
somewhere. Usage from node is trivial, and custom bundles fairly
straightforward, but the web is a mess. As always.