Added configuration to Schulze.
This includes adding tiebreak functionality, seat count, and logging.
Due to the fact that the ordering step is a fundamental part of this
method, _all_ tiebreaks are passed to the logger, even if they're not
relevant to the results. I may eventually look to change this.

The most significant change here is an alteration in the strength
measure used, which is now configurable as well and defaults to using
the margin of victory as strength. This is a change from the prior
version, which used the absolute number of winning votes as the
measure.