Added logging and custom tiebreaking to plurality.
This involves a fairly substantial change in the structure of the code, which is
now especially careful not to invoke tiebreaks unless they are relevant to the
results of the count. This also means that all ties are explicitly dealt with,
which was _not_ the case before.

Also includes tests.